Save annotations and predictions
- predictions are generated after training a neural network.  We save
  these separetly so we can compare to the labels.

- annotations are full images saved with the manual labels.  These are
  slightly different than 'labels'.  Labels are generated from ROIs
drawn on the "annotations".  The work flow is that we usually draw an
ROI and then draw labels inside the ROI.  Sometimes we might label but
not draw an ROI (without an ROI these manual annotations won't be used
for training, but we may want to still save them for later use).