Refactor model widgets
Instead of explicitly adding a widget full of parameters for each model,
use automatic discovery to find any class that derives from BaseClass
then parse the fields to add to the widget.  This is done mostly in
'param_widget'.  For each BaseModel class a 'param_widget' will be
created which will be connected to the model.  Thus explicit wiring code
to pass gui parameters to each model is no longer needed, and new models
can be added without modifying any other code.