Change features to classes in some places
In a couple of places 'features' was used to refer to storage of class
names. This was because Napari layers put class names in a feature
dictionary.  However now we are working on ML features, so features will
refer to the feature maps for ML algorithms.  Thus changing some of the
pre-existing variables to avoid confusion.