Allow Napari-hub to find this plugin....
I've been conflicted about this because this plugin while useful,
requires complicated DL dependencies and GPU drivers to be useful.  The
problem is it needs one or more of Stardist/Cellpose/SAM but doesn't
need all.  I don't to add all that (plus GPU drivers) as dependencies so
am leaving it up to the user to install the deep learning dependencies
they want to use.