Replace 'label_only' flag with enum...
The enum can represent three modes.  ALL, LABEL_ONLY, and DL_PIXEL_ONLY.
The first allows full functionality.  The second removes almost
everything from the GUI for the case where the user only wants to label
images (often the scenario is the user labels images with Napari, but
runs DL via a script or notebook).  The third is for DL approaches that
only require pixel labelling so the bounding box layer and ML pixel layers
are not needed, the scenario is the user only wants Cellpose, Stardist,
Pytorch Semantic Segmentation or other pixel based DL method.