Add blocksize as an auto parameter
If 'auto' is true then currently we don't use a blocksize (in the future
maybe add a more sophisticated method to calculate a blocksize, but for
now if auto unchecked it just means don't process in chunks).

If 'auto' is unchecked we use the block size specified in the gui.