Rework loading labels
Rework loading the labels so that it follows the exact same logic as
when saving them.  We construct the same by using the original file name
and appending an zero padded index to it.  We check if the name exists
and if so load the label and increment the index and check again.  This
way we can load multiple labels for an image.

The previous method just checked for all files that contained the image
name.  But this was flawed if one image had another images name within
it's name (ie cells.jpg and cells_control.jpg).