Add logic to handle different dataset splits
Data split can be 'train' or 'validation'.  In the future we may add
'test'.  The type of split will default to 'train'

The type of split can be changed by selecting the 'label box'
layer, selecting an roi and pressing the 'v' key to switch to
'validation'.  Press 'v' again and the type of split should toggle back
to 'train'.

'train':  Label (and downstream generated patches) will be used for deep
learning training.

'validation': Downstream generated patches will be saved in a separate
directory with 'validation' appended.  These will not be used for deep
learning training, but instead are intended to be used to assess
performance of the algorithm.