Allow tile size to be automatically set
This means that when harvested for the gui, tile size will have a 'auto'
check box.  If 'auto' is set tile size will be set to None (no tiling
will be used).

Todo:  May need to reword the 'auto' check box.  It's used to allow a
parameter to use the 'automatic' value instead of manually set.